Allow user to display all results
Allow print_distro_summaries() to take a hashref as argument.  If that hashref
has a 'cat_summaries' element that is true, then display all summaries on
STDOUT with appropriate header.  Document this, as well as internal method
print_distro_summary().  Test this option and use it in p5-dist-backcompat
(though we might consider making that a command-line switch).