Introduce category 'cpan_makefile_pl'
In etc/dist-backcompat-distro-metadata.txt, implement a column
'cpan_makefile_pl' where the recommended approach will be to provide the
distro with the Makefile.PL found in the most recent CPAN release.

Allow for this to be processed during init().

Threads-Queue and Threads-Semaphore are two such distros.  Until now, each is
failing during the 'configure' stage -- because they have no Makefile.PL in
core.