Don't hardcode paths to metadata files
Once this library is installed from CPAN, we won't be able to locate these
files automatically within the distro.  So we need to be able to pass them
explicitly to the constructor.

We'll retain the previously determined default values, but caution that they
are meaningful only within the distro itself (e.g., within t/*.t files).

TODO: Write tests of the front-end program p5-dist-backcompat which use
explicitly set values.